# src/swaggerDocument.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Library API",
        "version": "1.0.0",
        "description": "Simple blog for authenticated users to create post then edit or delete it"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ]
, "tags": [
    {
        "name": "Users",
        "description": "API for users registered in the system"
    }
],

"paths": {
    "/users": {
        "get": {
            "tags": [
                "Users"
            ],
            "summary": "Get all users in system",
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Users"
                    }
                },
                "400": {
                    "description": "can't get users"
                }
            }
        }
    },
     "/users/register" :{  
        "post": {
            "tags": [
                "Users"
            ],
            "summary": "Create a new user in system",
            "requestBody": {
                "description": "User Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }  
            },
            "produces": [
                "application/json"
            ],
            "responses": {
                "201": {
                    "description": "OK",
                    "schema": {
                        "User":{
                            "type": "User Object",
                            "required":{
                               "name":"",
                               "email":"",
                               "password":""
                   }}
                    }
                },
                "400": {
                    "description": "can't register"
                }
            }
            
        }
    },
    "/users/login": {
        "post": {
            "tags": [
                "Users"
            ],
            "summary": "Log In user",
            "requestBody": {
                "description": "Login Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Users"
                    }
                },
                "400": {
                    "description": "wrong crediential"
                }
            }
        }
    },
    "/users/posts":{
        "get": {
            "tags": [
                "Users"
            ],
            "summary": "Get all posts related to the authenricated user",
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Users"
                    }
                },
                "400": {
                    "description": "you haven't posts to show"
                },
                "401": {
                    "description": "you aren't authorized"
                }
            }
        }
    },
    "/users/posts/{id}":{
        "parameters": [
            {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "given post ID",
                "type": "number"
            }
        ],
        "get": {
            "tags": [
                "Users"
            ],
            "summary": "get one post off this id related to current user",
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "post with id",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Posts"
                    }
                },
                "400": {
                    "description": "post not found."
                }
            }
        },
        "patch": {
            "summary": "Update post with given ID related to logged user",
            "tags": [
                "Users"
            ],
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Post with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ],
            "requestBody": {
                "description": "Posts Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/Posts"
                        }
                    }
                }
            },
           
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Posts"
                    }
                },
                "400": {
                    "description": "post can't be updated"
                },
                "401": {
                    "description": "You aren't authorized to update post"
                }
            }
        },
        "delete": {
            "summary": "Delete post with given ID",
            "tags": [
                "Users"
            ],
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Delete post with id for current user",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ],
            "responses": {
                "204": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                },
                "400": {
                    "description": "Post not found to be deleted"
                }
            }
        }
    },
    "/posts":{
        "get": {
            "tags": [
                "Posts"
            ],
            "summary": "Get all posts in system",
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Posts"
                    }
                },
                "400": {
                    "description": "can't get posts"
                }
            }
        } 
    },
    "/posts/create":{
        "post": {
            "tags": [
                "Posts"
            ],
            "summary": "Authenticated user create a new post",
            "requestBody": {
                "description": "Post Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                }  
            },
            "produces": [
                "application/json"
            ],
            "responses": {
                "201": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Posts"
                        }
                    }
                },
                "400": {
                    "description": "Post coudn't be created"
                },
                "401": {
                    "description": "You aren't authorized"
                }
            }  
        }
    },
"definitions": {
        "id": {
            "properties": {
                "id": {
                    "type": "number"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Users": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "additionalProperties": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
        "Post": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "Posts": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    }
                }
        },
        "Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}